Add addCustomizer kotlin extension to FixtureMonkeyBuilder
