+++ fixture-monkey-kotlin/src/main/kotlin/com/navercorp/fixturemonkey/kotlin/FixtureMonkeyBuilderExtensions.kt
@@ -1,0 +1,27 @@
+/*
+ * Fixture Monkey
+ *
+ * Copyright (c) 2021-present NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.fixturemonkey.kotlin
+
+import com.navercorp.fixturemonkey.FixtureMonkeyBuilder
+import com.navercorp.fixturemonkey.kotlin.customizer.KArbitraryCustomizer
+import com.navercorp.fixturemonkey.kotlin.customizer.toArbitraryCustomizer
+
+inline fun <reified T : Any> FixtureMonkeyBuilder.addCustomizer(
+    customizer: KArbitraryCustomizer<T>
+): FixtureMonkeyBuilder = this.addCustomizer(T::class.java, customizer.toArbitraryCustomizer())
